Refactor package structure for ApiDashboard components
- Moved `ApiDashboardController`, `ApiDashboardService`, `ApiDashboardControllerInterface`, and `WorkingBalanceResponse` to `presentation.api.dashboard` package.
